Fix `as_strided` functionalization for lazy backend.
[ghstack-poisoned]
