+++ build_variables.bzl
@@ -993,6 +993,7 @@ aten_cpu_source_non_codegen_list = [
     "aten/src/ATen/ParallelOpenMP.cpp",
     "aten/src/ATen/ParallelThreadPoolNative.cpp",
     "aten/src/ATen/PythonTorchFunctionTLS.cpp",
+    "aten/src/ATen/RegisterFunctionalizeManual.cpp",
     "aten/src/ATen/ThreadLocalPythonObjects.cpp",
     "aten/src/ATen/ScalarOps.cpp",
     "aten/src/ATen/SparseTensorImpl.cpp",
